requirements.txt
celery
appdirs==1.4.3
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.7.0rc1
distlib==0.3.0
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
facebook-sdk==3.1.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.5.0
mysqlclient==2.0.1
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
SQLAlchemy==1.3.20
sqlparse==0.4.1
urllib3==1.25.10
utils==1.0.1
virtualenv==20.0.4
whitenoise==5.2.0
zipp==2.2.0
